requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
# Optional for Celery async processing:
# celery>=5.3.0
# redis>=5.0.0
